+++ src/main/kotlin/com/coder/gateway/sdk/CoderRestClientService.kt
@@ -13,17 +13,14 @@ import com.coder.gateway.sdk.v2.models.User
 import com.coder.gateway.sdk.v2.models.Workspace
 import com.coder.gateway.sdk.v2.models.WorkspaceAgent
 import com.coder.gateway.sdk.v2.models.WorkspaceBuild
+import com.coder.gateway.sdk.v2.models.WorkspaceTransition
 import com.google.gson.Gson
 import com.google.gson.GsonBuilder
 import com.intellij.openapi.components.Service
-import okhttp3.Cookie
-import okhttp3.HttpUrl.Companion.toHttpUrlOrNull
-import okhttp3.JavaNetCookieJar
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import java.net.CookieManager
 import java.net.HttpURLConnection.HTTP_CREATED
 import java.net.URL
 import java.time.Instant
@@ -42,26 +39,17 @@ class CoderRestClientService {
      * @throws [AuthenticationResponseException] if authentication failed.
      */
     fun initClientSession(url: URL, token: String): User {
-        val cookieUrl = url.toHttpUrlOrNull()!!
-        val cookieJar = JavaNetCookieJar(CookieManager()).apply {
-            saveFromResponse(
-                cookieUrl,
-                listOf(Cookie.parse(cookieUrl, "session_token=$token")!!)
-            )
-        }
         val gson: Gson = GsonBuilder()
             .registerTypeAdapter(Instant::class.java, InstantConverter())
             .setPrettyPrinting()
             .create()
 
-        val interceptor = HttpLoggingInterceptor()
-        interceptor.setLevel(HttpLoggingInterceptor.Level.BASIC)
         retroRestClient = Retrofit.Builder()
             .baseUrl(url.toString())
             .client(
                 OkHttpClient.Builder()
-                    .addInterceptor(interceptor)
-                    .cookieJar(cookieJar)
+                    .addInterceptor { it.proceed(it.request().newBuilder().addHeader("Coder-Session-Token", token).build()) }
+                    .addInterceptor(HttpLoggingInterceptor().apply { setLevel(HttpLoggingInterceptor.Level.BASIC) })
                     .build()
             )
             .addConverterFactory(GsonConverterFactory.create(gson))
@@ -91,7 +79,7 @@ class CoderRestClientService {
             throw WorkspaceResponseException("Could not retrieve Coder Workspaces:${workspacesResponse.code()}, reason: ${workspacesResponse.message()}")
         }
 
-        return workspacesResponse.body()!!
+        return workspacesResponse.body()!!.workspaces
     }
 
     private fun buildInfo(): BuildInfo {
@@ -126,7 +114,7 @@ class CoderRestClientService {
     }
 
     fun startWorkspace(workspaceID: UUID, workspaceName: String): WorkspaceBuild {
-        val buildRequest = CreateWorkspaceBuildRequest(null, "start", null, null, null)
+        val buildRequest = CreateWorkspaceBuildRequest(null, WorkspaceTransition.START, null, null, null, null)
         val buildResponse = retroRestClient.createWorkspaceBuild(workspaceID, buildRequest).execute()
         if (buildResponse.code() != HTTP_CREATED) {
             throw WorkspaceResponseException("Failed to build workspace ${workspaceName}: ${buildResponse.code()}, reason: ${buildResponse.message()}")
@@ -136,7 +124,7 @@ class CoderRestClientService {
     }
 
     fun stopWorkspace(workspaceID: UUID, workspaceName: String): WorkspaceBuild {
-        val buildRequest = CreateWorkspaceBuildRequest(null, "stop", null, null, null)
+        val buildRequest = CreateWorkspaceBuildRequest(null, WorkspaceTransition.STOP, null, null, null, null)
         val buildResponse = retroRestClient.createWorkspaceBuild(workspaceID, buildRequest).execute()
         if (buildResponse.code() != HTTP_CREATED) {
             throw WorkspaceResponseException("Failed to stop workspace ${workspaceName}: ${buildResponse.code()}, reason: ${buildResponse.message()}")
@@ -145,10 +133,10 @@ class CoderRestClientService {
         return buildResponse.body()!!
     }
 
-    fun updateWorkspace(workspaceID: UUID, workspaceName: String, lastWorkspaceTransition: String, templateID: UUID): WorkspaceBuild {
+    fun updateWorkspace(workspaceID: UUID, workspaceName: String, lastWorkspaceTransition: WorkspaceTransition, templateID: UUID): WorkspaceBuild {
         val template = template(templateID)
 
-        val buildRequest = CreateWorkspaceBuildRequest(template.activeVersionID, lastWorkspaceTransition, null, null, null)
+        val buildRequest = CreateWorkspaceBuildRequest(template.activeVersionID, lastWorkspaceTransition, null, null, null, null)
         val buildResponse = retroRestClient.createWorkspaceBuild(workspaceID, buildRequest).execute()
         if (buildResponse.code() != HTTP_CREATED) {
             throw WorkspaceResponseException("Failed to update workspace ${workspaceName}: ${buildResponse.code()}, reason: ${buildResponse.message()}")