+++ src/main/kotlin/com/coder/gateway/sdk/v2/models/Template.kt
@@ -10,12 +10,26 @@ data class Template(
     @SerializedName("updated_at") val updatedAt: Instant,
     @SerializedName("organization_id") val organizationIterator: UUID,
     @SerializedName("name") val name: String,
-    @SerializedName("provisioner") val provisioner: String,
+    @SerializedName("display_name") val displayName: String,
+    @SerializedName("provisioner") val provisioner: ProvisionerType,
     @SerializedName("active_version_id") val activeVersionID: UUID,
     @SerializedName("workspace_owner_count") val workspaceOwnerCount: Int,
+    @SerializedName("active_user_count") val activeUserCount: Int,
+    @SerializedName("build_time_stats") val buildTimeStats: Map<WorkspaceTransition, TransitionStats>,
     @SerializedName("description") val description: String,
-    @SerializedName("max_ttl_ms") val maxTTLMillis: Long,
-    @SerializedName("min_autostart_interval_ms") val minAutostartIntervalMillis: Long,
+    @SerializedName("icon") val icon: String,
+    @SerializedName("default_ttl_ms") val defaultTTLMillis: Long,
     @SerializedName("created_by_id") val createdByID: UUID,
     @SerializedName("created_by_name") val createdByName: String,
+    @SerializedName("allow_user_cancel_workspace_jobs") val allowUserCancelWorkspaceJobs: Boolean,
 )
+
+enum class ProvisionerType {
+    @SerializedName("echo")
+    ECHO,
+
+    @SerializedName("terraform")
+    TERRAFORM
+}
+
+data class TransitionStats(val p50: Long, val p95: Long)